refactor: final auth simplification
