+++ routes/auth.php
@@ -1,5 +1,1 @@
 <?php
-
-use App\Http\Controllers\Auth\AuthenticatedSessionController;
-use Illuminate\Support\Facades\Route;
-
